+++ CHANGELOG.md
@@ -4,6 +4,8 @@ UNRELEASED
 * [ [#267](https://github.com/digitalfabrik/lunes-cms/issues/267) ] Add professions in the vocabulary list
 * [ [#497](https://github.com/digitalfabrik/lunes-cms/issues/497) ] Mark words that have an image with icon in training sets view
 * [ [#477](https://github.com/digitalfabrik/lunes-cms/issues/477) ] Add definition field
+* [ [#516](https://github.com/digitalfabrik/lunes-cms/issues/516) ] Add fields for plural of alternative words and plural article
+
 
 2023.7.0
 --------